BBVAFSW3T3-26 | Agrego logica para autorizar deleteo segun rol y id de usuario

+++ wallet/src/main/java/com/bbva/wallet/controllers/UserController.java
@@ -1,7 +1,11 @@
 package com.bbva.wallet.controllers;
 
+import com.bbva.wallet.entities.User;
+import com.bbva.wallet.enums.RoleName;
 import com.bbva.wallet.services.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,9 +17,14 @@ import org.springframework.security.access.prepost.PreAuthorize;
 public class UserController {
     private final UserService userService;
 
+//    @PreAuthorize("hasRole('ADMIN')") todo: revisar, no filtra por rol ni por id
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasRole('ADMIN') or @userService.findById(#id).orElse(null)?.getEmail() == principal.username")
-    public void deleteById(@PathVariable Long id) {
+    public ResponseEntity<String> deleteById(@PathVariable Long id, Authentication authentication ) {
+        User userLoggedIn = (User) authentication.getPrincipal();
+        if( userLoggedIn.getId() != id && !(userLoggedIn.getAuthorities().contains(RoleName.ADMIN) )) {
+            return ResponseEntity.status(403).body("No tienes permisos para eliminar este usuario");
+        }
         userService.deleteById(id);
+        return ResponseEntity.noContent().build();
     }
 }
